requirements.txt
# ── Web framework ───────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# ── Git operations ──────────────────────────
gitpython>=3.1.43

# ── Docker sandbox ──────────────────────────
docker>=7.0.0

# ── LLM / AI agent ─────────────────────────
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=2.0.0
langchain-groq>=0.1.0
langchain-core>=0.2.0
langgraph>=0.0.40

# ── Settings ────────────────────────────────
pydantic>=2.6.0
pydantic-settings>=2.2.0

# ── Testing ─────────────────────────────────
pytest>=8.1.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
